fix: rename util.h to utility.h to avoid conflict
