+++ include/log.h
@@ -9,6 +9,8 @@
  */
 #ifndef __LOG_H__
 #define __LOG_H__
+#include <utility.h>
+
 #include <cstdint>
 #include <vector>
