+++ src/log.cpp
@@ -5,7 +5,7 @@
  */
 
 #include <log.h>
-#include <util.h>
+#include <utility.h>
 
 #include <string>
