+++ src/naive_kv.cpp
@@ -5,7 +5,7 @@
  */
 
 #include <naive_kv.h>
-#include <util.h>
+#include <utility.h>
 
 int NaiveKV::put(const std::string key, const std::string val)
 {
