+++ src/utility.cpp
@@ -2,4 +2,6 @@
  * This file implements any helper functions.
  */
 
-#include <util.h>
+#include <utility.h>
+
+
